user_service/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
requests
djangorestframework-simplejwt